Add user_account_setup()
In setup.py, I've added a method (user_account_setup()) that requests the user to set a password to secure their account and authenticate them to use PyPass. The username of the account is automatically set to the original username that was originally created from the operating system. Within the method, the Username and Password are saved to PyPassdb.sqlite3